# application/src/main/resources/static/openapi/canyoningApp-openapi.yml
openapi: 3.0.3
info:
  title: Swagger CanyoningApp - OpenAPI 3.0
  description: Swagger of canyoning app
  contact:
    email: delgadofernandez.luis@gmail.com
  version: 0.0.1
servers:
  - url: http://localhost:8080
  #- url: https://misbarrancosbdesarrollo.com:8443
...
paths:
...
  /reviews/canyon:
    post:
      tags:
        - Review
      summary: Create a new review
      description: Create a new review
      operationId: createCanyonReview
      requestBody:
        description: data from the review
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanyonReview'
        required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanyonReviewResponse'
        400:
          $ref: '#/components/responses/400BadRequest'
    get:
      tags:
        - Review
      summary: Get all review
      description: Get all review filtering by differents fields
      operationId: getCanyonReviews
      parameters:
        - name: canyonId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Id'
        - name: page
          in: query
          description: Número de página (empezando desde 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Número de elementos por página
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Criterios de ordenación en el formato propiedad(,asc|desc). El orden por defecto es ascendente.
          required: false
          schema:
            type: string
            default: "reviewId,asc"
      x-spring-paginated: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimpleCanyonReview'
          headers:
            Link:
              $ref: '#/components/headers/Link'
            X-Page-Number:
              $ref: '#/components/headers/X-Page-Number'
            X-Total-Elements:
              $ref: '#/components/headers/X-Total-Elements'
            X-Total-Pages:
              $ref: '#/components/headers/X-Total-Pages'
        400:
          $ref: '#/components/responses/400BadRequest'
  /reviews/canyon/{canyonReviewId}:
    get:
      tags:
        - Review
      summary: Get a review
      description: Get review by id
      operationId: getCanyonReview
      parameters:
        - name: canyonReviewId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanyonReviewResponse'
        400:
          $ref: '#/components/responses/400BadRequest'
    put:
      tags:
        - Review
      summary: Update your review
      description: Update your review
      operationId: updateCanyonReview
      parameters:
        - name: canyonReviewId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanyonReview'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanyonReviewResponse'
        400:
          $ref: '#/components/responses/400BadRequest'
    delete:
      tags:
        - Review
      summary: Deletes a review
      description: delete a review
      operationId: deleteCanyonReview
      parameters:
        - name: canyonReviewId
          in: path
          description: Review id to delete
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        204:
          description: 'Succesfull delete'
        400:
          $ref: '#/components/responses/400BadRequest'
  /reviews/canyon/{canyonReviewId}/share:
    get:
      tags:
        - Review
      summary: Share your review qr
      description: Share your review qr
      operationId: shareCanyonReview
      parameters:
        - name: canyonReviewId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        200:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
    post:
      tags:
        - Review
      summary: Insert an user on your review
      description: Insert an user on your review
      operationId: insertUserIntoCanyonReview
      parameters:
        - name: canyonReviewId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdsRequest'
      responses:
        200:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
...
  /messages:
    post:
      tags:
        - Message
      summary: Post a new message
      description: Post a new message
      operationId: postMessage
      requestBody:
        description: data from the message
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageIn'
        required: true
      responses:
        201:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
  /messages/{activityType}/activity/{activityId}:
    get:
      tags:
        - Message
      summary: Get all message about an activity
      description: Get all message about an activity
      operationId: getMessages
      parameters:
        - name: activityType
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ActivityType'
        - name: activityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - name: page
          in: query
          description: Número de página (empezando desde 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Número de elementos por página
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Criterios de ordenación en el formato propiedad(,asc|desc). El orden por defecto es messageId desc.
          required: false
          schema:
            type: string
            default: "messageId,desc"
      x-spring-paginated: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageOut'
        400:
          $ref: '#/components/responses/400BadRequest'
  /messages/{messageId}:
    put:
      tags:
        - Message
      summary: Update a meesage
      description: Update a message
      operationId: updateMessage
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        description: data from the message
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageIn'
        required: true
      responses:
        200:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
  /message/{messageId}/activityType/{activityType}:
    delete:
      tags:
        - Message
      summary: Delete a message
      description: Delete a message
      operationId: deleteMessage
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - name: activityType
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ActivityType'
      responses:
        204:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
...
components:
  schemas:
    UserIdsRequest:
      type: array
      items:
        $ref: '#/components/schemas/Id'
    Message:
      type: object
      properties:
        messageId:
          $ref: '#/components/schemas/Id'
        placeId:
          $ref: '#/components/schemas/Id'
        message:
          type: string
        date:
          type: string
          format: date-time
    MessageIn:
      allOf:
        - $ref: '#/components/schemas/Message'
        - properties:
            typePlace:
              $ref: '#/components/schemas/ActivityType'
    MessageOut:
      allOf:
        - $ref: '#/components/schemas/Message'
        - properties:
            user:
              $ref: '#/components/schemas/SimpleUser'
    ActivityType:
      type: string
      enum:
        - CANYON
...
    SimpleCanyon:
      type: object
      properties:
        canyonId:
          $ref: '#/components/schemas/Id'
        name:
          type: string
        season:
          type: string
        river:
          type: string
        population:
          type: string
...
    SimpleCanyonReview:
      properties:
        canyonReviewId:
          $ref: '#/components/schemas/Id'
        canyon:
          $ref: '#/components/schemas/SimpleCanyon'
        user:
          $ref: '#/components/schemas/SimpleUser'
    CanyonReview:
      allOf:
        - $ref: '#/components/schemas/SimpleCanyonReview'
        - properties:
            date:
              type: string
              format: date
            duration:
              type: integer
            combinedCar:
              type: boolean
            caudal:
              type: integer
            description:
              type: string
            users:
              type: array
              items:
                $ref: '#/components/schemas/SimpleUser'
    CanyonReviewResponse:
      allOf:
        - $ref: '#/components/schemas/CanyonReview'
        - properties:
            qr:
              type: string
    SimpleUser:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/Id'
        email:
          type: string
          example: theUser
        name:
          type: string
          example: John
        lastName:
          type: string
          example: James
        location:
          type: string
        description:
          type: string
        birthDay:
          type: string
          format: date
...
    Id:
      type: integer
      format: int64
    Error:
      type: object
      properties:
        errorCode:
          type: string
        description:
          type: string
        cause:
          type: string
  responses:
    400BadRequest:
      description: Bad Request. Check the parameters or body of the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    basic:
      type: http
      scheme: basic
  headers:
    Link:
      description: >-
        The Link header provides information about the pagination of the response.
      schema:
        type: string
    X-Page-Number:
      description: >-
        The current page number of the response.
      schema:
        type: integer
    X-Total-Elements:
      description: >-
        The total number of elements available in the response.
      schema:
        type: integer
    X-Total-Pages:
      description: >-
        The total number of pages available in the response.
      schema:
        type: integer
security:
  - basic: [ ]